requirements.txt
aif360==0.4.0
alembic==1.5.8
atomicwrites==1.4.0
attrdict==2.0.1
attrs==20.3.0
bravado==11.0.3
bravado-core==5.17.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cliff==3.7.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==1.5.0
colorama==0.4.4
colorlog==5.0.1
cycler==0.10.0
datatable==0.11.1
docopt==0.6.2
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
joblib==1.0.1
jsonpickle==1.5.2
jsonpointer==2.1
jsonref==0.2
jsonschema==3.2.0
keras-tuner==1.0.2
kiwisolver==1.3.1
lightgbm==3.2.1
llvmlite==0.36.0
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.1
memory-profiler==0.58.0
monotonic==1.6
msgpack==1.0.2
munch==2.5.0
neptune-client==0.9.5
neptune-contrib==0.27.1
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
optuna==2.7.0
packaging==20.9
pandas==1.2.4
pbr==5.5.1
Pillow==8.2.0
plotly==4.14.3
pluggy==0.13.1
prettytable==2.1.0
psutil==5.8.0
py==1.10.0
py-cpuinfo==8.0.0
pyaml==20.4.0
PyJWT==2.0.1
pyparsing==2.4.7
pyperclip==1.8.2
pyreadline3==3.3
pyrsistent==0.17.3
pytest==6.2.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3987==1.3.8
sacred==0.8.2
scikit-learn==0.24.1
scikit-optimize==0.8.1
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.1
shap==0.39.0
simplejson==3.17.2
six==1.15.0
sklearn==0.0
slicer==0.0.7
smmap==4.0.0
SQLAlchemy==1.4.9
stevedore==3.3.0
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
tabulate==0.8.9
tempeh==0.1.12
terminaltables==3.1.0
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webcolors==1.11.1
websocket-client==0.58.0
wrapt==1.12.1
xgboost==1.4.0
yellowbrick==1.3.post1
